Delete the default copy and move operations
For Basic instances, the default copy and move operations are not defined (by
default), so we explicitly remove them. Now code like this fails:

Integer i, j; // Fails
i = j; // Fails (even if the previous line succeeded)

So the user is forced to use RCP<Integer> for copies as intented.