Travis: Add -Werror to CXXFLAGS
This will make the Travis builds fail if there is any warning, which is what we
want, as we should treat warnings as errors during development. This change
only affects Travis --- in the default builds, the warnings are printed but not
treated as errors, so that if the user is using a newer compiler that triggers
some warnings that we didn't previously see in development, the build still
succeeds.