Introduce WITH_SYMENGINE_CHECKS
This option turns on all Debug time checks (Teuchos and other asserts). The
advantage is that you can turn this option independently of the Release/Debug
options. By default, Release turns it off and Debug turns it on, but you can
override it, which is useful for e.g. debugging some production code in Release
mode that segfaults (one keeps Release mode but turns on
WITH_SYMENGINE_CHECKS), or for testing that the code compiles both with and
without WITH_SYMENGINE_CHECKS (so one can keep Debug mode for fast
compilation).