Fix CMake default options
Fixes #1422, https://github.com/symengine/symengine.py/issues/206.

CMake assumes the following for the default Release options:

* Must work with any code
* The same binary must work on any platform

This essentialy leads just to `-O2`. More aggresive options like `-O3` or
`-ffast-math` cannot be used because this options is considered "unsafe" and
some codes miscompile with it. Similarly, `-march=native` cannot be used
because then the binary will only work on the platform for which it was
compiled.

Our default options assume the following for "our default" Release options:

* Get the best performance on the given platform
* Works for any code that is written in such a robust way so that it works
  correctly when these aggressive compiler options are used. It is not
  SymEngine specific.

This leads to `-O3 -march=native -ffast-math -funroll-loops` for gcc/clang.
Such options work for a wide variety of codes that are written in a way that
they allow the compiler the freedom to optimize aggressively.

These default options can be overriden by the user in any way he wishes by
setting the CMAKE_CXX_FLAGS_[RELEASE,DEBUG] variables. For example
distributions can set CMAKE_CXX_FLAGS_RELEASE="-O3 -ffast-math -funroll-loops",
or something like that in order for the binary to work on all platforms. The
idea is that no matter how CMAKE_CXX_FLAGS_RELEASE is modified, the code will
always build. The C++11 options do not belong into `cmake/UserOverride.cmake`
because some codes do not need them, and some codes require them and will fail
if the user modifies the defaults --- such options belong into CMakeLists.txt.

Project specific flags must be set in CMakeLists.txt. For SymEngine, since we
depend on C++11, we must set the C++11 options in CMakeLists.txt. It is up to
us how configurable we make this for the user. Since our code fails to build
without the C++11 enabled, we do not need to have it configurable (but we could
if we wanted).

Specifically, for homebrew, which sets its own CMAKE_CXX_FLAGS_RELEASE flags,
this patch makes SymEngine work just fine, because now the user can set any
flags in CMAKE_CXX_FLAGS_RELEASE and we always append the essential project
specific options in CMakeLists.txt (in our case C++11 options).