Support multi-precision incomplete gamma function
MPFR 4.0.0 has been out for awhile now so it would be nice to make use
of more of its special functions. This does so for UpperGamma and once
it is in place, it will be trivial to do the same for LowerGamma. This
currently puts the new code directly in functions.cpp since there is
only a unified evaluate interface for OneArgFunction.

Signed-off-by: Connor Behan <connor.behan@gmail.com>