cmake: fix pthread detection
The homebrew FindPTHREAD.cmake module was hardcoding the path to libpthread
in the installed CMake config file which broke consumers when upgrading from
< glibc-2.34 (e.g. glibc-2.33 -> glibc-2.34).

CMake provides FindThreads which does the job fine for us. Tested with both
glibc-2.33 (with a previously bad generated file, and then fixed) and glibc-2.35.

This should be fine on musl and friends too. This brings symengine in line
with most CMake packages.

Bug: https://bugs.gentoo.org/849803
Fixes: https://github.com/symengine/symengine/issues/1910