Fix linking to LLVM dylib
Detect if LLVM has been built with the request to link to dylib,
and use dylib rather than trying to link components libs in that case.
This fixes undesirable static linking when static libraries are
installed, and an error when they are not.

Note that this probably is not the best way of handling the problem
(upstream recommends using something like `llvm_config(symengine
USE_SHARED ${SYMENGINE_LLVM_COMPONENTS})`) but it makes minimal changes
to the current complex logic, and therefore is least likely to break
something.

Fixes #2074