Fix LLVM dylib use in installed CMake config
This is a followup to #2075 that fixes linking against LLVM dylib
in reverse dependencies of SymEngine (e.g. the Python package).
Unfortunately, my original patch missed this second occurrence.

Given that we permit overriding the dylib use via a cache variable
and it is unlikely for a dylib install to suddenly become non-dylib
let's store the value in the config file.

Signed-off-by: Michał Górny <mgorny@gentoo.org>